% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{plot.tvarGIRF}
\alias{plot.tvarGIRF}
\title{Plot the GIRFs}
\usage{
\method{plot}{tvarGIRF}(x, y, ...)
}
\arguments{
\item{x}{The result of a call to `GIRF`}

\item{y}{Not used}

\item{...}{Additional arguments; not used}
}
\description{
Creates a ggplot of your responses, with each response in its own panel
}
