% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{print.tvarGIRF}
\alias{print.tvarGIRF}
\title{Print the GIRFs}
\usage{
\method{print}{tvarGIRF}(x, ...)
}
\arguments{
\item{x}{The result of a call to `GIRF`}

\item{...}{Additional arguments; not used}
}
\description{
Print the GIRFs
}
