% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{summary.tvarGIRF}
\alias{summary.tvarGIRF}
\title{Print a summary of some estimated GIRFs}
\usage{
\method{summary}{tvarGIRF}(object, ...)
}
\arguments{
\item{object}{The result of a call to `GIRF`}

\item{...}{Additional arguments; not used}
}
\description{
Print a summary of some estimated GIRFs
}
