% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{tidy.tvarGIRF}
\alias{tidy.tvarGIRF}
\title{Convert GIRF responses to tidy form}
\usage{
tidy.tvarGIRF(x, ...)
}
\arguments{
\item{x}{GIRFs returned by a call to `GIRF`}

\item{...}{Additional arguments to tidying method; not used}
}
\value{
A tibble in long form
}
\description{
Convert GIRF responses to tidy form
}
