% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{View.tvarGIRF}
\alias{View.tvarGIRF}
\title{View GIRFs}
\usage{
View.tvarGIRF(x, title)
}
\arguments{
\item{x}{The result of a call to `GIRF`}

\item{title}{title for viewer window. Defaults to name of tvar if missing.}
}
\description{
Invoke a spreadsheet-style data view on some estimates GIRFs
}
