% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIRF.R
\name{GIRF}
\alias{GIRF}
\title{Generalised impulse response function for a threshold VAR}
\usage{
GIRF(tvar, shock, horizon = 20, H = 200, R = 500, restrict.to = NA)
}
\arguments{
\item{tvar}{An estimated tvar (from the tsDyn package) for which you wish to calculate a GIRF.}

\item{shock}{A column specifying which (reduced form) shock you wish to impose.}

\item{horizon}{(default 20) How many periods to simulate after the shock?}

\item{H}{(default 200) How many histories to sample.}

\item{R}{(default 500) How many times to replicate the forward simulation for each history.}

\item{restrict.to}{(integer) Do you want to restrict to a particular regime in the shock period (histories that lead to starting in any other regime will be ignored).}
}
\description{
This function constructs GIRFs for an threshold VAR (from the tsDyn package) using variation of the method from Koop, Pesaran and Potter (1996).
}
\details{
Specifically, the method bootstraps a history and then simulates the TVAR forward.
First it does so by imposing that the shock in the first period is equal to the supplied (reduced form) shock. Shocks in all periods subsequent are
bootstrapped from the distribution of reduced form shocks, _conditional on the current regime_. It repeats the simulation a second time, but with
a bootstrapped shock in place of the imposed shock (and then randomly bootstrapping shocks to simulate the TVAR as before). Since the TVAR is allowed
to endogenously change regime, the shock series will differ between the two simulations - both because each is randomly selected, and because the two
simulations will not necessarily have the same path for the regime. The difference between the two simulations is then saved as the response of
each of the variables in the TVAR to the shock.

This process is repeated a number of times for each history, and for a number of histories and then averaged to construct the GIRF.

Importantly, the supplied shock (and the shocks that are sampled) are reduced form. This means you are responsible for supplying a sensibly identifited
initial shock. But after that, by using reduced form shocks, there is no need to take a stand about identifying structural shocks for all of the monte
carlo simulation periods.

Creates a (potentially multipanel) time series graph from a ts objecct. Supports bar and line (and combinations of).
}
\examples{


}
